requirements.txt
fake-useragent
pygame
selenium
discord.py
python-dotenv
colorama
asyncio
tinydb